requirements.txt
discord_interactions
PyNaCl
gTTS
PyYAML
boto3==1.28.54